requirements.txt
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
